refactor : 상속 없애기 / LottoBall은 Int만 받을 수 있도록 변경 / 수익률 표기 변경

+++ src/main/kotlin/LottoApplication.kt
@@ -19,9 +19,9 @@ fun main() {
 
     val ballNums = StringUtil.split(InputView.getResultLottoBallNums(), ",")
     val resultLottoBalls = LottoResultBallBundle(
-        LottoBallBundle(ballNums.map { LottoBall(it) })
+        LottoBallBundle(ballNums.map { LottoBall(it.toInt()) })
     )
-    val bonusWinningNumber = LottoResultBonusBall(LottoBall(InputView.getBonusNumber()))
+    val bonusWinningNumber = LottoResultBonusBall(LottoBall(InputView.getBonusNumber().toInt()))
 
     val lottoResult = LottoResult(gambler.getTickets(), resultLottoBalls, bonusWinningNumber)
